backend/anti-cheat/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
prisma
bson
numpy
pandas
pyarrow
scikit-learn
matplotlib
shap
keras
keras-tuner
tensorflow-cpu
